Moved Newman execution to worker thread to help with memory leaks
